fix post request on save video button

+++ controllers/videos.js
@@ -1,24 +1,29 @@
 const express = require('express');
 const router = express.Router();
 const db = require('../models');
+const auth = require('../middleware/auth');
 
-//ROUTE TO GET A SINGLE VIDEO FROM THE DATABASE
-router.get('/Video', function(req, res) {
-	db.Video.find({})
+//Get all videos
+router.get('/', auth, (req, res) => {
+	db.Video.find({ user: req.user })
 		.then(dbVideos => res.json(dbVideos))
-		.catch(err => res.status(502).json(err));
+		.catch(err => res.json(err));
 });
 
-router.post('/', (req, res) => {
-	db.Video.create(req.body)
+//Create video
+router.post('/', auth, (req, res) => {
+	const newVideo = { user: req.user, ...req.body };
+	db.Video.create(newVideo)
 		.then(dbVideos => res.json(dbVideos))
-		.catch(err => res.status(502).json(err));
+		.catch(err => res.json(err));
 });
-
-router.delete('/:id', function(req, res) {
+router.delete('/:id', auth, (req, res) => {
+	// Make sure user delete only their own video
+	if (dbVideos.user.toString() !== req.user)
+		return res.status(401).json({ msg: 'Not authorized' });
 	db.Video.findByIdAndDelete(req.params.id)
 		.then(dbVideo => res.json(dbVideo))
-		.catch(err => res.status(502).json(err));
+		.catch(err => res.json(err));
 });
 
 module.exports = router;
